refactor(misc): refine directory structure
