+++ src/config.ts
@@ -9,10 +9,7 @@ if (!NODE_ENV) {
 }
 
 // loads environment variables from .env.<NODE_ENV> file in project root folder
-const envFile = path.join(
-  path.resolve(__dirname, "../../"),
-  `.env.${NODE_ENV}`,
-);
+const envFile = path.join(path.resolve(__dirname, "../"), `.env.${NODE_ENV}`);
 
 if (fs.existsSync(envFile)) {
   dotenv.config({
