+++ src/routes/index.ts
@@ -1,0 +1,3 @@
+// router들을 한 곳에 모아서 export해주는 역할을 합니다.
+export { default as scrapResultRouter } from "./scrapResult";
+export { default as testRouter } from "./test";
